✨ Feat: 연차 추가 Employee 엔티티 변경, EmpOff 관련 머지 주의

+++ src/main/java/com/insadong/application/off/dto/OffDTO.java
@@ -16,7 +16,7 @@ public class OffDTO {
 
 	private LocalDate offEnd;					// 종료일
 
-	private Long offDay;					// 일수
+	private Double offDay;					// 일수
 
 	private String signReason;				// 신청 사유
